Use node:test and node:assert
This moves the actual test suite to node:test.  A small refactor was done so that each entry in the test suite gets a test name.

Also the two assertions at the end of the test file were leftover from when abbrev had a monkeyPatch() method.  They have been removed.